Changed Functors/Monads/State folder topology
