+++ src/Functors/Monads/State/index.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * state monad functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/EvalState.php';
+require_once __DIR__ . '/ExecState.php';
+require_once __DIR__ . '/Get.php';
+require_once __DIR__ . '/Gets.php';
+require_once __DIR__ . '/Modify.php';
+require_once __DIR__ . '/Put.php';
+require_once __DIR__ . '/RunState.php';
+require_once __DIR__ . '/State.php';
